agents/data_retrieval/requirements.txt
# Data Retrieval Agent Dependencies
#
# NOTE: All dependencies are now installed globally in the parent directory.
# This file is kept for reference and documentation purposes.
#
# To install all dependencies for the multi-agent system:
#   cd /Users/nchua/Desktop/Cursor
#   pip install -r requirements.txt
#
# Individual agent requirements (already satisfied globally):

# Core dependencies (INSTALLED GLOBALLY)
pandas>=2.0.0
PyYAML>=6.0
python-dateutil>=2.8.2

# Snowflake connectivity (INSTALLED GLOBALLY)
snowflake-connector-python>=3.0.0

# Testing dependencies (INSTALLED GLOBALLY)
pytest>=7.0.0
pytest-cov>=4.0.0

# Enhanced functionality (INSTALLED GLOBALLY)
jinja2>=3.1.0  # For advanced query templating
requests>=2.28.0  # For HTTP clients